ai_cartoon_generator/requirements.txt
# Core Django Framework
Django>=4.2,<5.0

# PostgreSQL Database Adapter
psycopg2-binary>=2.9.9

# Image Processing
Pillow>=10.1.0

# Google OAuth Authentication
django-allauth>=0.57.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Environment Variables Management
python-decouple>=3.8

# Security and Production
gunicorn>=21.2.0
whitenoise>=6.6.0
